@@ -1,4 +1,4 @@
-class UniqueIndexesForUsersTeams < ActiveRecord::Migration
+class UniqueIndexesForUsersTeams < ActiveRecord::Migration[4.2]
   def up
     remove_index :teams, :name
     add_index :teams, :name, unique: true
