@@ -1,4 +1,4 @@
-class AddDisplayToTeams < ActiveRecord::Migration
+class AddDisplayToTeams < ActiveRecord::Migration[4.2]
   def change
     add_column :teams, :display, :string
   end
