@@ -1,4 +1,4 @@
-class CreateUploads < ActiveRecord::Migration[5.0]
+class CreateUploads < ActiveRecord::Migration[4.2][5.0]
   def change
     create_table :uploads do |t|
       t.belongs_to :user, foreign_key: true
