@@ -1,4 +1,4 @@
-class AddFileFingerprintToUploads < ActiveRecord::Migration[5.0]
+class AddFileFingerprintToUploads < ActiveRecord::Migration[4.2][5.0]
   def change
     add_column :uploads, :file_fingerprint, :binary
   end
