@@ -1,4 +1,4 @@
-class AddVisibleToUploads < ActiveRecord::Migration[5.0]
+class AddVisibleToUploads < ActiveRecord::Migration[4.2][5.0]
   def change
     add_column :uploads, :visible, :boolean
   end
